experimental: Move some code to test_utils.go
