+++ pkg/loadbalancer/experimental/bpf_reconciler_test.go
@@ -5,11 +5,8 @@ package experimental
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"net/netip"
 	"slices"
-	"strconv"
 	"strings"
 	"testing"
 
@@ -37,63 +34,6 @@ const (
 	LoadBalancer  = loadbalancer.SVCTypeLoadBalancer
 	LocalRedirect = loadbalancer.SVCTypeLocalRedirect
 )
-
-var (
-	// special addresses that are replaced by the test runner.
-	autoAddr = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("0.0.0.1"),
-		L4Addr:      loadbalancer.L4Addr{},
-		Scope:       0,
-	}
-	zeroAddr = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("0.0.0.3"),
-		L4Addr:      loadbalancer.L4Addr{},
-		Scope:       0,
-	}
-
-	extraFrontend = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("10.0.0.2"),
-		L4Addr: loadbalancer.L4Addr{
-			Protocol: loadbalancer.TCP,
-			Port:     80,
-		},
-		Scope: 0,
-	}
-
-	// backend addresses
-	backend1 = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("10.1.0.1"),
-		L4Addr: loadbalancer.L4Addr{
-			Protocol: loadbalancer.TCP,
-			Port:     80,
-		},
-		Scope: 0,
-	}
-	backend2 = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("10.1.0.2"),
-		L4Addr: loadbalancer.L4Addr{
-			Protocol: loadbalancer.TCP,
-			Port:     80,
-		},
-		Scope: 0,
-	}
-
-	// frontendAddrs are assigned to the <auto>/autoAddr. Each test set is run with
-	// each of these.
-	frontendAddrs = []loadbalancer.L3n4Addr{
-		parseAddrPort("10.0.0.1:80"),
-		parseAddrPort("[2001::1]:80"),
-	}
-
-	nodePortAddrs = []netip.Addr{
-		netip.MustParseAddr("10.0.0.3"),
-		netip.MustParseAddr("2002::1"),
-	}
-)
-
-// mapDump is a dump of a BPF map. These are generated by the dump() method, which
-// solely defines the format.
-type mapDump = string
 
 type testCase struct {
 	name string
@@ -795,27 +735,6 @@ var testCases = [][]testCase{
 	sessionAffinityTestCases,
 }
 
-func parseAddrPort(s string) loadbalancer.L3n4Addr {
-	addrS, portS, found := strings.Cut(s, "]:")
-	if found {
-		// IPv6
-		addrS = addrS[1:] // drop [
-	} else {
-		// IPv4
-		addrS, portS, found = strings.Cut(s, ":")
-		if !found {
-			panic("bad <ip:port>")
-		}
-	}
-	addr := types.MustParseAddrCluster(addrS)
-	port, _ := strconv.ParseInt(portS, 10, 16)
-	return *loadbalancer.NewL3n4Addr(
-		loadbalancer.TCP,
-		addr, uint16(port), loadbalancer.ScopeExternal,
-	)
-
-}
-
 func TestBPFOps(t *testing.T) {
 
 	lc := hivetest.Lifecycle(t)
@@ -926,141 +845,6 @@ func showMaps(m []mapDump) string {
 	}
 	w.WriteString("},\n")
 	return w.String()
-}
-
-type numeric interface {
-	~int | ~uint32 | ~uint16
-}
-
-// TODO: Figure out what to do about the IDs. If we want to do fault inject the
-// operations will be retried and the ID allocations are non-deterministic.
-func sanitizeID[Num numeric](n Num, sanitize bool) string {
-	if !sanitize {
-		return strconv.FormatInt(int64(n), 10)
-	}
-	if n == 0 {
-		return "<zero>"
-	}
-	return "<non-zero>"
-}
-
-// dump the load-balancing maps into a concise format for assertions in tests.
-func dump(lbmaps lbmaps, feAddr loadbalancer.L3n4Addr, sanitizeIDs bool) (out []mapDump) {
-	out = []string{}
-
-	replaceAddr := func(addr net.IP, port uint16) (s string) {
-		if addr.IsUnspecified() {
-			s = "<zero>"
-			return
-		}
-		switch addr.String() {
-		case feAddr.AddrCluster.String():
-			s = "<auto>"
-		case nodePortAddrs[0].String():
-			s = "<nodePort>"
-		case nodePortAddrs[1].String():
-			s = "<nodePort>"
-		default:
-			s = addr.String()
-			if addr.To4() == nil {
-				s = "[" + s + "]"
-			}
-			s = fmt.Sprintf("%s:%d", s, port)
-		}
-		return
-	}
-
-	svcCB := func(svcKey lbmap.ServiceKey, svcValue lbmap.ServiceValue) {
-		svcKey = svcKey.ToHost()
-		svcValue = svcValue.ToHost()
-		addr := svcKey.GetAddress()
-		addrS := replaceAddr(addr, svcKey.GetPort())
-		if svcKey.GetScope() == loadbalancer.ScopeInternal {
-			addrS += "/i"
-		}
-		out = append(out, fmt.Sprintf("SVC: ID=%s ADDR=%s SLOT=%d BEID=%s COUNT=%d QCOUNT=%d FLAGS=%s",
-			sanitizeID(svcValue.GetRevNat(), sanitizeIDs),
-			addrS,
-			svcKey.GetBackendSlot(),
-			sanitizeID(svcValue.GetBackendID(), sanitizeIDs),
-			svcValue.GetCount(),
-			svcValue.GetQCount(),
-			strings.ReplaceAll(
-				loadbalancer.ServiceFlags(svcValue.GetFlags()).String(),
-				", ", "+"),
-		))
-	}
-	if err := lbmaps.DumpService(svcCB); err != nil {
-		panic(err)
-	}
-
-	beCB := func(beKey lbmap.BackendKey, beValue lbmap.BackendValue) {
-		beValue = beValue.ToHost()
-		addr := beValue.GetAddress()
-		addrS := addr.String()
-		if addr.To4() == nil {
-			addrS = "[" + addrS + "]"
-		}
-		stateS, _ := loadbalancer.GetBackendStateFromFlags(beValue.GetFlags()).String()
-		out = append(out, fmt.Sprintf("BE: ID=%s ADDR=%s:%d STATE=%s",
-			sanitizeID(beKey.GetID(), sanitizeIDs),
-			addrS,
-			beValue.GetPort(),
-			stateS,
-		))
-	}
-	if err := lbmaps.DumpBackend(beCB); err != nil {
-		panic(err)
-	}
-
-	revCB := func(revKey lbmap.RevNatKey, revValue lbmap.RevNatValue) {
-		revKey = revKey.ToHost()
-		revValue = revValue.ToHost()
-
-		var addr string
-
-		switch v := revValue.(type) {
-		case *lbmap.RevNat4Value:
-			addr = replaceAddr(v.Address.IP(), v.Port)
-
-		case *lbmap.RevNat6Value:
-			addr = replaceAddr(v.Address.IP(), v.Port)
-		}
-
-		out = append(out, fmt.Sprintf("REV: ID=%s ADDR=%s",
-			sanitizeID(revKey.GetKey(), sanitizeIDs),
-			addr,
-		))
-	}
-	if err := lbmaps.DumpRevNat(revCB); err != nil {
-		panic(err)
-	}
-
-	affCB := func(affKey *lbmap.AffinityMatchKey, _ *lbmap.AffinityMatchValue) {
-		affKey = affKey.ToHost()
-		out = append(out, fmt.Sprintf("AFF: ID=%s BEID=%d",
-			sanitizeID(affKey.RevNATID, sanitizeIDs),
-			affKey.BackendID,
-		))
-	}
-
-	if err := lbmaps.DumpAffinityMatch(affCB); err != nil {
-		panic(err)
-	}
-
-	srcRangeCB := func(key lbmap.SourceRangeKey, _ *lbmap.SourceRangeValue) {
-		key = key.ToHost()
-		out = append(out, fmt.Sprintf("SRCRANGE: ID=%s CIDR=%s",
-			sanitizeID(key.GetRevNATID(), sanitizeIDs),
-			key.GetCIDR(),
-		))
-	}
-	if err := lbmaps.DumpSourceRange(srcRangeCB); err != nil {
-		panic(err)
-	}
-
-	slices.Sort(out)
-	return
 }
 
 type mapKeyValue struct {
