+++ pkg/loadbalancer/experimental/k8s_test.go
@@ -4,12 +4,9 @@
 package experimental
 
 import (
-	"bytes"
 	"context"
 	"os"
 	"path"
-	"regexp"
-	"slices"
 	"strings"
 	"testing"
 	"time"
@@ -19,12 +16,8 @@ import (
 	"github.com/cilium/statedb"
 	"github.com/cilium/statedb/reconciler"
 	"github.com/cilium/stream"
-	"github.com/pmezard/go-difflib/difflib"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8sRuntime "k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
 
 	"github.com/cilium/cilium/pkg/datapath/tables"
 	"github.com/cilium/cilium/pkg/hive"
@@ -32,59 +25,9 @@ import (
 	"github.com/cilium/cilium/pkg/k8s/resource"
 	slim_corev1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
 	slim_discovery_v1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/discovery/v1"
-	slim_fake "github.com/cilium/cilium/pkg/k8s/slim/k8s/client/clientset/versioned/fake"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/testutils"
 )
-
-var (
-	slimDecoder k8sRuntime.Decoder
-)
-
-func init() {
-	slimScheme := k8sRuntime.NewScheme()
-	slim_fake.AddToScheme(slimScheme)
-	slimScheme.AddKnownTypes(slim_corev1.SchemeGroupVersion, &metav1.List{})
-	slimDecoder = serializer.NewCodecFactory(slimScheme).UniversalDeserializer()
-}
-
-func decodeObject[Obj k8sRuntime.Object](t *testing.T, file string) Obj {
-	bytes, err := os.ReadFile(file)
-	require.NoError(t, err, "ReadFile(%s)", file)
-	obj, _, err := slimDecoder.Decode(bytes, nil, nil)
-	require.NoError(t, err, "Decode(%s)", file)
-	return obj.(Obj)
-}
-
-func readObjects[Obj k8sRuntime.Object](t *testing.T, dataDir string, prefix string) (out []Obj) {
-	ents, err := os.ReadDir(dataDir)
-	require.NoError(t, err, "ReadDir(%s)", dataDir)
-
-	for _, ent := range ents {
-		if strings.HasPrefix(ent.Name(), prefix) && strings.HasSuffix(ent.Name(), ".yaml") {
-			out = append(out, decodeObject[Obj](t, path.Join(dataDir, ent.Name())))
-		}
-	}
-	return
-}
-
-func upsertEvent[Obj k8sRuntime.Object](obj Obj) resource.Event[Obj] {
-	return resource.Event[Obj]{
-		Object: obj,
-		Key:    resource.NewKey(obj),
-		Kind:   resource.Upsert,
-		Done:   func(error) {},
-	}
-}
-
-func deleteEvent[Obj k8sRuntime.Object](obj Obj) resource.Event[Obj] {
-	return resource.Event[Obj]{
-		Object: obj,
-		Key:    resource.NewKey(obj),
-		Kind:   resource.Delete,
-		Done:   func(error) {},
-	}
-}
 
 func TestIntegrationK8s(t *testing.T) {
 	t.Parallel()
@@ -380,73 +323,6 @@ func testIntegrationK8s(t *testing.T, testDataPath string) {
 		5*time.Second,
 		50*time.Millisecond) {
 		t.Fatalf("Expected BPF maps to be empty after pruning, instead they contain: %v", lastDump)
-	}
-}
-
-// sanitizeTables clears non-deterministic data in the table output such as timestamps.
-func sanitizeTables(dump []byte) []byte {
-	r := regexp.MustCompile(`\([^\)]* ago\)`)
-	return r.ReplaceAll(dump, []byte("(??? ago)"))
-}
-
-func checkTablesAndMaps(db *statedb.DB, writer *Writer, maps lbmaps, testDataPath string) bool {
-	allDone := true
-	count := 0
-	for fe := range writer.Frontends().All(db.ReadTxn()) {
-		if fe.Status.Kind != reconciler.StatusKindDone {
-			allDone = false
-		}
-		count++
-	}
-	if count == 0 || !allDone {
-		return false
-	}
-
-	var tableBuf bytes.Buffer
-	writer.DebugDump(db.ReadTxn(), &tableBuf)
-	actualTables := tableBuf.Bytes()
-
-	var expectedTables []byte
-	if expectedData, err := os.ReadFile(path.Join(testDataPath, "expected.tables")); err == nil {
-		expectedTables = expectedData
-	}
-	actualTables = sanitizeTables(actualTables)
-	expectedTables = sanitizeTables(expectedTables)
-
-	os.WriteFile(path.Join(testDataPath, "actual.tables"), actualTables, 0644)
-
-	var expectedMaps []mapDump
-	if expectedData, err := os.ReadFile(path.Join(testDataPath, "expected.maps")); err == nil {
-		expectedMaps = strings.Split(strings.TrimSpace(string(expectedData)), "\n")
-	}
-	actualMaps := dump(maps, frontendAddrs[0], true)
-
-	actualPath := path.Join(testDataPath, "actual.maps")
-	os.WriteFile(
-		actualPath,
-		[]byte(strings.Join(actualMaps, "\n")+"\n"),
-		0644,
-	)
-	return bytes.Equal(actualTables, expectedTables) && slices.Equal(expectedMaps, actualMaps)
-}
-
-func logDiff(t *testing.T, fileA, fileB string) {
-	t.Helper()
-
-	contentsA, err := os.ReadFile(fileA)
-	require.NoError(t, err)
-	contentsB, _ := os.ReadFile(fileB)
-
-	diff := difflib.UnifiedDiff{
-		A:        difflib.SplitLines(string(contentsA)),
-		B:        difflib.SplitLines(string(contentsB)),
-		FromFile: fileA,
-		ToFile:   fileB,
-		Context:  2,
-	}
-	text, _ := difflib.GetUnifiedDiffString(diff)
-	if len(text) > 0 {
-		t.Logf("\n%s", text)
 	}
 }
